Add Sign Up Page and Update Navigation
Added a new sign-up page and updated navigation routes to include the sign-up screen. Also, modified the login screen to navigate to the sign-up page when the user clicks on "إنشاء حسابك الآن". Additionally, updated the onboard screen to include a button for creating a new account and navigating to the sign-up page.