Add Shorebird configuration and update entitlements
This commit introduces a new Shorebird configuration file
(shorebird.yaml) to manage app updates and includes the app ID
for identification. Additionally, it updates the macOS
entitlements to allow network client access, which is necessary
for Shorebird's functionality. The main.dart file is also
modified to include a Shorebird patch comment for versioning.
Finally, the pubspec.yaml is updated to include the new
shorebird.yaml in the assets.