Add UploadBookCubit and refactor imports
This commit introduces the `UploadBookCubit` to handle the
uploading of book data, including cover images, to Supabase.
It also refactors the import statements in `biblio.dart` and
`add_book.dart` to correctly reference the new cubit structure.
Additionally, new states for the upload process are defined
in `upload_book_state.dart`.