Add FetchLocatedBooksCubit and Update Book Page
Implemented FetchLocatedBooksCubit to fetch books based on the user's location.
This includes state management for loading, success, and error states.
Updated the book page to display user-related information and enhanced the UI for better user experience.

Also, modified the NewBooksListview to utilize the new cubit forfetching and displaying books, ensuring a responsive and dynamic interface.